Fix #37: standardize error handling and call the C destructors (#41)
Description
This is a stabilization PR, aiming to fix some of the memory issues we have with this SDK:

The static Kuzzle exception throwing method has been changed: it should not deallocate the memory of the structure provided to it. It now only throws the appropriate exception.
A new KUZZLE_API macro has been added: it DRYifies and standardizes the way the Kuzzle API is requested, and how errors are handled. The main purpose of that macro is to correctly free the C result struct before throwing an exception
All controller actions now use the new KUZZLE_API macro
Fix some controller actions using delete to free the C result struct, instead of the corresponding C destructor
Note: there are other obvious memleaks remaining. I will address those in other, dedicated PRs, to make this stabilization pass easier to review.

Other changes
As per the changes made in the C SDK: the user_rights_result struct does not contain a NULL-terminated result pointer anymore, but a rights pointer, with its size stored in the rights_length property
Fix non-standard indentation in a few files