Fix #40 (memleak prevention): return a User C++ class instead of a kuzzle_user* pointer (#44)
:warning: depends on https://github.com/kuzzleio/sdk-cpp/pull/42 :warning: 
:warning: depends on https://github.com/kuzzleio/sdk-c/pull/47 :warning: 

# Description

The `kuzzle_free_user_result` destructor was incomplete, and didn't free its `kuzzle_user` property. This was made on purpose, or so it seems, but this is unclear, and so can (and even will) lead to memory leaks.

I saw 2 possible solutions for that one, additional to fixing the dtor to make it consistent with the others:
1. add another, internal dtor for the `user_result` struct, leaving the kuzzle_user pointer alone, and use it in internal functions
2. wrap the `user_result` struct in a C++ class and return that instead of the raw pointer

I found the second solution clearer and less prone to confusion in the long run.