[bugfix] do not invoke cgo functions before cgo initializes (deadlock) (#59)
# Description

Using the C++ SDK in its current state results in programs hanging up while loading the SDK library.
The hangup is caused by the `defaultOptions` static, invoking the `Options` default constructor which, in turn, uses CGo.
Since that variable is static, it is initialized during the library load, before CGo has a chance to finish initializing its threads and mutexes. This results in a deadlock (see this related issue: https://github.com/golang/go/issues/26184 )

This PR fixes that issue by removing the need of a static "default options" variable. 
To do so, we need the Options `c_opts` method, which converts the C++ Options class into its equivalent `options` C struct, to be constant. This allows the use of delegating constructors without the need of a pre-allocated default Options instance.