Add missing "isReady" protocol bridge (#72)
# Description

As the title says: a new protocol has been added recently in the Go and Cgo layers, but not in the C++ one. 

When invoking certain methods such as the Kuzzle `playQueue` one, a crash occurs because the Cgo layer tries to invoke the non-existant C++ bridge.

# How to test

Run the `playQueue` documentation test: it crashes without this fix, and it passes with it.